feat: implement selecting submission

+++ backend/src/main/java/com/quicklance/backend/dto/Task.java
@@ -2,4 +2,4 @@ package com.quicklance.backend.dto;
 
 import java.time.LocalDate;
 
-public record Task(Long id, String title, String description, Integer credits, User author, LocalDate createdAt) {}
+public record Task(Long id, String title, String description, Long credits, User author, LocalDate createdAt) {}
